refactor: integration tests directories
