+++ tests/integration/contract_examples/mocks/football_prediction_market_get_contract_schema_for_code.py
