+++ tests/integration/contract_examples/mocks/storage_get_contract_schema_for_code.py
