+++ tests/integration/contract_examples/mocks/user_storage_get_contract_schema_for_code.py
