+++ tests/integration/contract_examples/mocks/wizard_get_contract_schema_for_code.py
