+++ tests/integration/contract_examples/test_football_prediction_market.py
@@ -1,4 +1,4 @@
-# tests/e2e/test_wizard_of_coin.py
+# tests/e2e/test_storage.py
 
 from tests.common.request import (
     deploy_intelligent_contract,
@@ -6,10 +6,10 @@ from tests.common.request import (
     payload,
     post_request_localhost,
 )
-from tests.integration.mocks.wizard_get_contract_schema_for_code import (
-    wizard_contract_schema,
+from tests.integration.contract_examples.mocks.football_prediction_market_get_contract_schema_for_code import (
+    football_prediction_market_contract_schema,
 )
-from tests.integration.mocks.call_contract_function import (
+from tests.integration.contract_examples.mocks.call_contract_function import (
     call_contract_function_response,
 )
 
@@ -22,11 +22,10 @@ from tests.common.response import (
 from tests.common.accounts import create_new_account
 
 
-def test_wizard_of_coin():
-    print("test_wizard_of_coin")
-    # Validators
+def test_football_prediction_market():
+    # Validators Setup
     result = post_request_localhost(
-        payload("create_random_validators", 10, 8, 12, ["openai"], None, "gpt-4o-mini")
+        payload("create_random_validators", 5, 8, 12, ["openai"], None, "gpt-4o")
     ).json()
     assert has_success_status(result)
 
@@ -34,32 +33,36 @@ def test_wizard_of_coin():
     from_account = create_new_account()
 
     # Get contract schema
-    contract_code = open("examples/contracts/wizard_of_coin.py", "r").read()
+    contract_code = open("examples/contracts/football_prediction_market.py", "r").read()
     result_schema = post_request_localhost(
         payload("get_contract_schema_for_code", contract_code)
     ).json()
     assert has_success_status(result_schema)
-    assert_dict_exact(result_schema, wizard_contract_schema)
+    assert_dict_exact(result_schema, football_prediction_market_contract_schema)
 
     # Deploy Contract
     call_method_response_deploy, transaction_response_deploy = (
         deploy_intelligent_contract(
-            from_account, contract_code, f'{{"have_coin": true}}'
+            from_account,
+            contract_code,
+            f'{{"game_date": "2024-06-26", "team1": "Georgia", "team2": "Portugal"}}',
         )
     )
     assert has_success_status(transaction_response_deploy)
     contract_address = call_method_response_deploy["result"]["data"]["contract_address"]
 
-    # Call Contract Function
+    ########################################
+    ############# RESOLVE match ############
+    ########################################
     _, transaction_response_call_1 = call_contract_method(
         from_account,
         contract_address,
-        "ask_for_coin",
-        ["Can you please give me my coin?"],
+        "resolve",
+        [],
     )
     assert has_success_status(transaction_response_call_1)
 
-    # Assert format
+    # Assert response format
     assert_dict_struct(transaction_response_call_1, call_contract_function_response)
 
     delete_validators_result = post_request_localhost(
