+++ tests/integration/contract_examples/test_storage.py
@@ -6,10 +6,10 @@ from tests.common.request import (
     payload,
     post_request_localhost,
 )
-from tests.integration.mocks.football_prediction_market_get_contract_schema_for_code import (
-    football_prediction_market_contract_schema,
+from tests.integration.contract_examples.mocks.storage_get_contract_schema_for_code import (
+    storage_contract_schema,
 )
-from tests.integration.mocks.call_contract_function import (
+from tests.integration.contract_examples.mocks.call_contract_function import (
     call_contract_function_response,
 )
 
@@ -21,11 +21,14 @@ from tests.common.response import (
 
 from tests.common.accounts import create_new_account
 
+INITIAL_STATE = "a"
+UPDATED_STATE = "b"
 
-def test_football_prediction_market():
+
+def test_storage():
     # Validators Setup
     result = post_request_localhost(
-        payload("create_random_validators", 5, 8, 12, ["openai"], None, "gpt-4o")
+        payload("create_random_validators", 10, 8, 12, ["openai"], None, "gpt-4o-mini")
     ).json()
     assert has_success_status(result)
 
@@ -33,37 +36,46 @@ def test_football_prediction_market():
     from_account = create_new_account()
 
     # Get contract schema
-    contract_code = open("examples/contracts/football_prediction_market.py", "r").read()
+    contract_code = open("examples/contracts/storage.py", "r").read()
     result_schema = post_request_localhost(
         payload("get_contract_schema_for_code", contract_code)
     ).json()
     assert has_success_status(result_schema)
-    assert_dict_exact(result_schema, football_prediction_market_contract_schema)
+    assert_dict_exact(result_schema, storage_contract_schema)
 
     # Deploy Contract
     call_method_response_deploy, transaction_response_deploy = (
         deploy_intelligent_contract(
-            from_account,
-            contract_code,
-            f'{{"game_date": "2024-06-26", "team1": "Georgia", "team2": "Portugal"}}',
+            from_account, contract_code, f'{{"initial_storage": "{INITIAL_STATE}"}}'
         )
     )
+
     assert has_success_status(transaction_response_deploy)
     contract_address = call_method_response_deploy["result"]["data"]["contract_address"]
 
-    ########################################
-    ############# RESOLVE match ############
-    ########################################
+    # Get Initial State
+    contract_state_1 = post_request_localhost(
+        payload("get_contract_state", contract_address, "get_storage", [])
+    ).json()
+    assert has_success_status(contract_state_1)
+    assert contract_state_1["result"]["data"] == INITIAL_STATE
+
+    # Update State
     _, transaction_response_call_1 = call_contract_method(
-        from_account,
-        contract_address,
-        "resolve",
-        [],
+        from_account, contract_address, "update_storage", [UPDATED_STATE]
     )
+
     assert has_success_status(transaction_response_call_1)
 
     # Assert response format
     assert_dict_struct(transaction_response_call_1, call_contract_function_response)
+
+    # Get Updated State
+    contract_state_2 = post_request_localhost(
+        payload("get_contract_state", contract_address, "get_storage", [])
+    ).json()
+    assert has_success_status(contract_state_2)
+    assert contract_state_2["result"]["data"] == UPDATED_STATE
 
     delete_validators_result = post_request_localhost(
         payload("delete_all_validators")
