+++ tests/integration/contract_examples/test_wizard_of_coin.py
@@ -1,4 +1,4 @@
-# tests/e2e/test_storage.py
+# tests/e2e/test_wizard_of_coin.py
 
 from tests.common.request import (
     deploy_intelligent_contract,
@@ -6,10 +6,10 @@ from tests.common.request import (
     payload,
     post_request_localhost,
 )
-from tests.integration.mocks.storage_get_contract_schema_for_code import (
-    storage_contract_schema,
+from tests.integration.contract_examples.mocks.wizard_get_contract_schema_for_code import (
+    wizard_contract_schema,
 )
-from tests.integration.mocks.call_contract_function import (
+from tests.integration.contract_examples.mocks.call_contract_function import (
     call_contract_function_response,
 )
 
@@ -21,12 +21,10 @@ from tests.common.response import (
 
 from tests.common.accounts import create_new_account
 
-INITIAL_STATE = "a"
-UPDATED_STATE = "b"
 
-
-def test_storage():
-    # Validators Setup
+def test_wizard_of_coin():
+    print("test_wizard_of_coin")
+    # Validators
     result = post_request_localhost(
         payload("create_random_validators", 10, 8, 12, ["openai"], None, "gpt-4o-mini")
     ).json()
@@ -36,46 +34,33 @@ def test_storage():
     from_account = create_new_account()
 
     # Get contract schema
-    contract_code = open("examples/contracts/storage.py", "r").read()
+    contract_code = open("examples/contracts/wizard_of_coin.py", "r").read()
     result_schema = post_request_localhost(
         payload("get_contract_schema_for_code", contract_code)
     ).json()
     assert has_success_status(result_schema)
-    assert_dict_exact(result_schema, storage_contract_schema)
+    assert_dict_exact(result_schema, wizard_contract_schema)
 
     # Deploy Contract
     call_method_response_deploy, transaction_response_deploy = (
         deploy_intelligent_contract(
-            from_account, contract_code, f'{{"initial_storage": "{INITIAL_STATE}"}}'
+            from_account, contract_code, f'{{"have_coin": true}}'
         )
     )
-
     assert has_success_status(transaction_response_deploy)
     contract_address = call_method_response_deploy["result"]["data"]["contract_address"]
 
-    # Get Initial State
-    contract_state_1 = post_request_localhost(
-        payload("get_contract_state", contract_address, "get_storage", [])
-    ).json()
-    assert has_success_status(contract_state_1)
-    assert contract_state_1["result"]["data"] == INITIAL_STATE
-
-    # Update State
+    # Call Contract Function
     _, transaction_response_call_1 = call_contract_method(
-        from_account, contract_address, "update_storage", [UPDATED_STATE]
+        from_account,
+        contract_address,
+        "ask_for_coin",
+        ["Can you please give me my coin?"],
     )
-
     assert has_success_status(transaction_response_call_1)
 
-    # Assert response format
+    # Assert format
     assert_dict_struct(transaction_response_call_1, call_contract_function_response)
-
-    # Get Updated State
-    contract_state_2 = post_request_localhost(
-        payload("get_contract_state", contract_address, "get_storage", [])
-    ).json()
-    assert has_success_status(contract_state_2)
-    assert contract_state_2["result"]["data"] == UPDATED_STATE
 
     delete_validators_result = post_request_localhost(
         payload("delete_all_validators")
